Add Relacionamento entre Classes e Módulo Tema

+++ src/postagem/services/postagem.service.ts
@@ -1,6 +1,7 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, ILike, Repository } from 'typeorm';
+import { TemaService } from '../../tema/services/tema.service';
 import { Postagem } from '../entities/postagem.entity';
 
 @Injectable()
@@ -8,16 +9,24 @@ export class PostagemService {
   constructor(
     @InjectRepository(Postagem)
     private postagemRepository: Repository<Postagem>,
+    private temaService: TemaService,
   ) {}
 
   async findAll(): Promise<Postagem[]> {
-    return await this.postagemRepository.find();
+    return await this.postagemRepository.find({
+      relations: {
+        tema: true,
+      },
+    });
   }
 
   async findById(id: number): Promise<Postagem> {
     const postagem = await this.postagemRepository.findOne({
       where: {
         id,
+      },
+      relations: {
+        tema: true,
       },
     });
 
@@ -27,25 +36,54 @@ export class PostagemService {
     return postagem;
   }
 
-  async findAllByTitulo(titulo: string): Promise<Postagem[]> {
+  async findByTitulo(titulo: string): Promise<Postagem[]> {
     return await this.postagemRepository.find({
       where: {
         titulo: ILike(`%${titulo}%`),
+      },
+      relations: {
+        tema: true,
       },
     });
   }
 
   async create(postagem: Postagem): Promise<Postagem> {
+    if (postagem.tema) {
+      const tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema)
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+
+      return await this.postagemRepository.save(postagem);
+    }
+
     return await this.postagemRepository.save(postagem);
   }
 
   async update(postagem: Postagem): Promise<Postagem> {
-    await this.findById(postagem.id);
+    const buscaPostagem: Postagem = await this.findById(postagem.id);
+
+    if (!buscaPostagem || !postagem.id)
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+
+    if (postagem.tema) {
+      const tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema)
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+
+      return await this.postagemRepository.save(postagem);
+    }
+
     return await this.postagemRepository.save(postagem);
   }
 
   async delete(id: number): Promise<DeleteResult> {
-    await this.findById(id);
+    const buscaPostagem = await this.findById(id);
+
+    if (!buscaPostagem)
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+
     return await this.postagemRepository.delete(id);
   }
 }
